otxdata-producer/requirements.txt
kafka-python
OTXv2
